Correção do id_acolhimento de animal ao registrar acolhimento

+++ src/main/java/sosanimais/com/example/app/model/entity/Animal.java
@@ -6,13 +6,13 @@ import sosanimais.com.example.app.model.AnimalInformacao;
 public class Animal {
 
     private Long id;
-    private int idBaia;
-    private int idAcolhimento;
+    private Long idBaia;
+    private Long idAcolhimento;
 
     private AnimalInformacao informacao;
 
 
-    public Animal(Long id, int idBaia, int idAcolhimento, AnimalInformacao informacao){
+    public Animal(Long id, Long idBaia, Long idAcolhimento, AnimalInformacao informacao){
             this.id = id;
             this.idBaia = idBaia;
             this.idAcolhimento = idAcolhimento;
@@ -28,19 +28,19 @@ public class Animal {
             this.id = id;
         }
 
-        public int getIdBaia () {
+        public Long getIdBaia () {
             return idBaia;
         }
 
-        public void setIdBaia ( int idBaia){
+        public void setIdBaia ( Long idBaia){
             this.idBaia = idBaia;
         }
 
-        public int getIdAcolhimento () {
+        public Long getIdAcolhimento () {
             return idAcolhimento;
         }
 
-        public void setIdAcolhimento ( int idAcolhimento){
+        public void setIdAcolhimento ( Long idAcolhimento){
             this.idAcolhimento = idAcolhimento;
         }
